feat: implement thumbnail show/hide functionality in ThumbnailBoxComponent and remove ShowThumbnailService
